feat: Scheduled jobs have their own module now. Logic moved to service.
