+++ adapters/schedules/src/test/java/be/xplore/pricescraper/ScrapeItemsJobTests.java
@@ -1,0 +1,17 @@
+package be.xplore.pricescraper;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+class ScrapeItemsJobTests {
+  private final ScrapeItemsJob scrapeItemsJob;
+
+  public ScrapeItemsJobTests() {
+    this.scrapeItemsJob = new ScrapeItemsJob(null);
+  }
+
+  @Test
+  void constructed() {
+    Assertions.assertNotNull(scrapeItemsJob);
+  }
+}
