+++ application/pom.xml
@@ -54,6 +54,12 @@
         </dependency>
         <dependency>
             <groupId>be.xplore</groupId>
+            <artifactId>schedules</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>be.xplore</groupId>
             <artifactId>localdummyscraper</artifactId>
             <version>0.0.1-SNAPSHOT</version>
             <scope>compile</scope>
