--- application/src/test/java/be/xplore/pricescraper/schedules/ScrapeItemsJobTests.java
@@ -1,18 +1,0 @@
-package be.xplore.pricescraper.schedules;
-
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import org.junit.jupiter.api.Test;
-
-class ScrapeItemsJobTests {
-  private final ScrapeItemsJob scrapeItemsJob;
-
-  public ScrapeItemsJobTests() {
-    this.scrapeItemsJob = new ScrapeItemsJob(null, null);
-  }
-
-  @Test
-  void constructed() {
-    assertNotNull(scrapeItemsJob);
-  }
-}
