+++ pom.xml
@@ -19,6 +19,7 @@
         <module>core</module>
         <module>adapters</module>
         <module>tests</module>
+        <module>adapters/schedules</module>
     </modules>
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
